refactor: share document text helpers
