+++ scripts/docs-site/llms-full.mjs
@@ -3,6 +3,7 @@ import crypto from "node:crypto";
 import fs from "node:fs";
 import path from "node:path";
 
+import { stripMdxForLlms, firstHeading, titleize, fileSlug } from "./document-text.mjs";
 import { ignoredDocDirs, ignoredDocFiles, localeLabels } from "./config.mjs";
 
 const root = process.cwd();
@@ -122,21 +123,6 @@ function entryFor(key, file) {
   };
 }
 
-function stripMdxForLlms(input) {
-  return input
-    .replace(/^import\s+.+?;?\s*$/gm, "")
-    .replace(/<([A-Z][A-Za-z0-9_.-]*)([^>]*)\/>/g, (_, name, attrs) => componentLabel(name, attrs))
-    .replace(/<([A-Z][A-Za-z0-9_.-]*)([^>]*)>/g, (_, name, attrs) => componentLabel(name, attrs))
-    .replace(/<\/[A-Z][A-Za-z0-9_.-]*>/g, "")
-    .replace(/\n{3,}/g, "\n\n");
-}
-
-function componentLabel(name, attrs) {
-  const parsed = Object.fromEntries([...String(attrs).matchAll(/([A-Za-z0-9_-]+)=(?:"([^"]*)"|'([^']*)')/g)].map((match) => [match[1], match[2] ?? match[3] ?? ""]));
-  const label = parsed.title ?? parsed.name ?? parsed.href ?? "";
-  return label ? `\n${label}\n` : `\n${name}\n`;
-}
-
 function docsOrigin() {
   return (canonicalOrigin || "https://docs.openclaw.ai").replace(/\/$/, "");
 }
@@ -164,38 +150,4 @@ function sourceMetadata() {
     sha: data.sha ?? data.sources?.openclaw?.sha ?? null,
     syncedAt: data.syncedAt ?? null,
   };
-}
-
-function fileSlug(rel) {
-  return normalizeSlug(rel.replace(/\.(md|mdx)$/, ""));
-}
-
-function normalizeSlug(value) {
-  return value.replace(/\/index$/, "") || "index";
-}
-
-function firstHeading(markdown) {
-  const heading = markdown.match(/^#\s+(.+)$/m)?.[1];
-  return heading === undefined ? undefined : textFromHtml(heading).trim();
-}
-
-function titleize(value) {
-  return value.replaceAll("-", " ").replace(/\b\w/g, (m) => m.toUpperCase());
-}
-
-function textFromHtml(value) {
-  let text = "";
-  let inTag = false;
-  for (const char of String(value)) {
-    if (char === "<") {
-      inTag = true;
-      continue;
-    }
-    if (char === ">") {
-      inTag = false;
-      continue;
-    }
-    if (!inTag) text += char;
-  }
-  return text;
 }
